feat: refactor & implement different gallery layout
