+++ src/lib/components/gallery/Gallery.helpers.ts
@@ -3,9 +3,9 @@ import type { Size } from "./Gallery.types";
 /**
  * Generates grid spans from image dimentions and grid's cell size.
  */
-export function getItemsSpan({ width, height }: Size, cellSize = 400): string {
+export function getMasonryItemSpan({ width, height }: Size, scale = 400): string {
 	return `
-		--column-end: span ${Math.floor(width / cellSize)};
-		--row-end: span ${Math.floor(height / cellSize)};
+		--masonry-item-column: span ${Math.floor(width / scale)};
+		--masonry-item-row: span ${Math.floor(height / scale)};
 	`;
 }
