+++ src/lib/components/gallery/gallery-image.module.css
@@ -1,0 +1,62 @@
+/* ovverride subcomponents styling */
+.link {
+	display: flex;
+	width: 100%;
+	height: 100%;
+	border-radius: var(--radius-l);
+}
+
+.figure,
+.figure picture {
+	width: 100%;
+	height: 100%;
+}
+
+.figure picture {
+	display: flex;
+}
+
+.figure {
+	position: relative;
+	display: flex;
+	overflow: hidden;
+	border-radius: var(--radius-l);
+	transition: border-color 0.25s ease;
+}
+
+.figure figcaption {
+	position: absolute;
+	bottom: 0;
+	color: white;
+	z-index: 1;
+	transition: transform 0.25s ease-in-out;
+	background-image: linear-gradient(transparent, rgb(109 84 199));
+	padding: var(--space-s);
+	padding-top: var(--space-l);
+	width: 100%;
+	font-family: var(--font-secondary);
+	transform: translateY(100%);
+}
+
+.figure img {
+	object-fit: cover;
+	background-color: var(--surface-2);
+	transition: transform 0.35s ease-in-out;
+}
+
+.figure:hover figcaption,
+.link:focus-visible figcaption {
+	transform: translateY(0);
+}
+
+/* state */
+
+@media (hover: hover) {
+	.figure img {
+		cursor: pointer;
+	}
+
+	.figure img:not(:global(.error)):hover {
+		transform: scale(1.1);
+	}
+}
