+++ src/lib/components/gallery/gallery.module.css
@@ -1,99 +1,10 @@
-/* container */
-
-.gallery {
-	justify-self: center;
-
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(var(--cell-size, 100px), 1fr));
-	grid-template-rows: repeat(auto-fit, var(--cell-size, 100px));
-	grid-auto-flow: dense;
-	place-items: stretch;
-	gap: calc(1.5 * var(--space-s));
-	width: 100%;
-}
-
-/* gallery-item */
-
-.gallery-item {
-	grid-row-end: var(--row-end);
-	grid-column-end: var(--column-end);
-	height: 100%;
-	position: relative;
-}
-
-/* image container */
-
-.figure,
-.figure picture {
-	width: 100%;
-	height: 100%;
-}
-
-.figure picture {
+.container {
 	display: flex;
+	flex-flow: column nowrap;
+	gap: var(--space-m);
 }
 
-.figure {
-	position: relative;
-	display: flex;
-	overflow: hidden;
-	border-radius: var(--radius-l);
-	transition: border-color 0.25s ease;
-}
-
-.figure figcaption {
-	position: absolute;
-	bottom: 0;
-	color: white;
-	z-index: 1;
-	transition: transform 0.25s ease-in-out;
-	background-image: linear-gradient(transparent, rgb(109 84 199));
-	padding: var(--space-s);
-	padding-top: var(--space-l);
-	width: 100%;
-	font-family: var(--font-secondary);
-	transform: translateY(100%);
-}
-
-.figure img {
-	object-fit: cover;
-	background-color: var(--surface-2);
-	transition: transform 0.35s ease-in-out;
-}
-
-.figure:hover figcaption {
-	transform: translateY(0);
-}
-
-@media screen and (max-width: 599px) {
-	.gallery {
-		grid-template-columns: 1fr;
-		grid-template-rows: unset;
-		grid-auto-columns: auto;
-	}
-
-	.gallery-item {
-		grid-column-end: unset;
-	}
-}
-
-/* state */
-
-@media (hover: hover) {
-	.gallery img {
-		cursor: pointer;
-	}
-
-	.gallery-item img:not(:global(.error)):hover {
-		transform: scale(1.1);
-	}
-}
-
-/* ovverride subcomponents styling */
-
-.gallery-item a {
-	display: flex;
-	width: 100%;
-	height: 100%;
-	border-radius: var(--radius-l);
+.control-panel > label {
+	width: min-content;
+	grid-auto-flow: column;
 }
