+++ src/lib/components/gallery/index.ts
@@ -1,2 +1,3 @@
 export { default as Gallery } from "./Gallery.svelte";
+export { default as GalleryGrid } from "./Grid.svelte";
 export type { GalleryItem } from "./Gallery.types";
